feat: new menu in sidebar
