+++ .prettierrc
@@ -5,6 +5,7 @@
   "printWidth": 100,
   "eslintIntegration": true,
   "endOfLine": "auto",
+  "trailingComma": "all",
   "parser": "typescript",
   "useTabs": false,
   "bracketSpacing": true,
