+++ src/app.feature/sideMenu/components/MenuIcon.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import styled from 'styled-components'
 import Image from 'next/image'
-import { MenuType } from '@app.modules/types/sideMenu'
+import { MenuType } from '@app.feature/sideMenu/types/sideMenu'
 
 type MenuIconProps = {
   path: string
