+++ src/app.feature/sideMenu/types/sideMenu.ts
@@ -1,0 +1,1 @@
+export type MenuType = 'Conjunctions' | 'Launch Conjunctions' | 'Watcher Catcher' | 'Favorites'
