+++ src/app.modules/hooks/useModal.ts
@@ -1,9 +1,9 @@
-import { modalType } from '@app.modules/types/modal'
+import { ModalType } from '@app.modules/types/modal'
 import { useDispatch, useSelector } from 'react-redux'
 import { RootState } from 'src/app.store/config/configureStore'
 import { closeModal, setModal } from 'src/app.store/modalStore/store.modalApp'
 
-export const useModal = (type: modalType) => {
+export const useModal = (type: ModalType) => {
   const dispatch = useDispatch()
   const { modalType, modalVisible } = useSelector((state: RootState) => state.modal)
