+++ src/app.modules/types/modal.ts
@@ -1,8 +1,7 @@
-export type modalType = 'LOGIN' | 'CONJUNCTIONS' | 'LAUNCHCONJUNCTIONS' | 'WATCHERCATCHER' | null
-
-export const modalTypeData = {
-  LOGIN: 'LOGIN',
-  CONJUNCTIONS: 'CONJUNCTIONS',
-  LAUNCHCONJUNCTIONS: 'LAUNCHCONJUNCTIONS',
-  WATCHERCATCHER: 'WATCHERCATCHER',
-}
+export type ModalType =
+  | 'LOGIN'
+  | 'CONJUNCTIONS'
+  | 'LAUNCHCONJUNCTIONS'
+  | 'WATCHERCATCHER'
+  | 'FAVORITES'
+  | null
