--- src/app.modules/types/sideMenu.ts
@@ -1,1 +1,0 @@
-export type MenuType = 'Conjunctions' | 'Launch Conjunctions' | 'Watcher Catcher'
